Move SystemRole code to its own package
